+++ Changelog.md
@@ -10,3 +10,4 @@ Bugfixes:
 * In setlist-based song statistics, only count title once per setlist (#102)
 * Button color standardization in song view mode switch (#99)
 * Standardized and extended additional information that can be shown on slides: Song Books, Author, Publisher, Copyright and CCLI ID (see Program Settings dialog -> Master formatting). (#100, #101)
+* Double-click on song in song statistics dialog opens song in editor (#103)